+++ src/filterTypes.js
@@ -1,79 +1,93 @@
-export const text = (rows, id, filterValue) => {
+export const text = (rows, ids, filterValue) => {
   rows = rows.filter(row => {
-    const rowValue = row.values[id]
-    return String(rowValue)
-      .toLowerCase()
-      .includes(String(filterValue).toLowerCase())
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return String(rowValue)
+        .toLowerCase()
+        .includes(String(filterValue).toLowerCase())
+    })
   })
   return rows
 }
 
 text.autoRemove = val => !val
 
-export const exactText = (rows, id, filterValue) => {
+export const exactText = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return rowValue !== undefined
-      ? String(rowValue).toLowerCase() === String(filterValue).toLowerCase()
-      : true
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return rowValue !== undefined
+        ? String(rowValue).toLowerCase() === String(filterValue).toLowerCase()
+        : true
+    })
   })
 }
 
 exactText.autoRemove = val => !val
 
-export const exactTextCase = (rows, id, filterValue) => {
+export const exactTextCase = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return rowValue !== undefined
-      ? String(rowValue) === String(filterValue)
-      : true
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return rowValue !== undefined
+        ? String(rowValue) === String(filterValue)
+        : true
+    })
   })
 }
 
 exactTextCase.autoRemove = val => !val
 
-export const includes = (rows, id, filterValue) => {
+export const includes = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return filterValue.includes(rowValue)
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return filterValue.includes(rowValue)
+    })
   })
 }
 
 includes.autoRemove = val => !val || !val.length
 
-export const includesAll = (rows, id, filterValue) => {
+export const includesAll = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return (
-      rowValue &&
-      rowValue.length &&
-      filterValue.every(val => rowValue.includes(val))
-    )
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return (
+        rowValue &&
+        rowValue.length &&
+        filterValue.every(val => rowValue.includes(val))
+      )
+    })
   })
 }
 
 includesAll.autoRemove = val => !val || !val.length
 
-export const exact = (rows, id, filterValue) => {
+export const exact = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return rowValue === filterValue
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return rowValue === filterValue
+    })
   })
 }
 
 exact.autoRemove = val => typeof val === 'undefined'
 
-export const equals = (rows, id, filterValue) => {
+export const equals = (rows, ids, filterValue) => {
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    // eslint-disable-next-line eqeqeq
-    return rowValue == filterValue
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      // eslint-disable-next-line eqeqeq
+      return rowValue == filterValue
+    })
   })
 }
 
 equals.autoRemove = val => val == null
 
-export const between = (rows, id, filterValue) => {
+export const between = (rows, ids, filterValue) => {
   let [min, max] = filterValue || []
 
   min = typeof min === 'number' ? min : -Infinity
@@ -86,8 +100,10 @@ export const between = (rows, id, filterValue) => {
   }
 
   return rows.filter(row => {
-    const rowValue = row.values[id]
-    return rowValue >= min && rowValue <= max
+    return ids.some(id => {
+      const rowValue = row.values[id]
+      return rowValue >= min && rowValue <= max
+    })
   })
 }